--- a/Godeps/_workspace/src/github.com/ugorji/go/codec/codecgen/z.go
+++ b/Godeps/_workspace/src/github.com/ugorji/go/codec/codecgen/z.go
@@ -1,0 +1,3 @@
+package main
+
+const genCodecPath = "github.com/ugorji/go/codec"