--- a/Godeps/_workspace/src/github.com/ugorji/go/codec/prebuild.go
+++ b/Godeps/_workspace/src/github.com/ugorji/go/codec/prebuild.go
@@ -1,0 +1,3 @@
+package codec
+
+//go:generate bash prebuild.sh