… Godeps/Godeps.json
… Godeps/_workspace/src/github.com/ugorji/go/codec/README.md
… Godeps/_workspace/src/github.com/ugorji/go/codec/binc.go
… Godeps/_workspace/src/github.com/ugorji/go/codec/cbor_test.go
… Godeps/_workspace/src/github.com/ugorji/go/codec/codec_test.go
… Godeps/_workspace/src/github.com/ugorji/go/codec/codecgen/README.md
… Godeps/_workspace/src/github.com/ugorji/go/codec/codecgen_test.go
… Godeps/_workspace/src/github.com/ugorji/go/codec/decode.go
… Godeps/_workspace/src/github.com/ugorji/go/codec/encode.go
… Godeps/_workspace/src/github.com/ugorji/go/codec/fast-path.generated.go
… Godeps/_workspace/src/github.com/ugorji/go/codec/fast-path.go.tmpl
… Godeps/_workspace/src/github.com/ugorji/go/codec/gen-dec-array.go.tmpl
… Godeps/_workspace/src/github.com/ugorji/go/codec/gen-dec-map.go.tmpl
… Godeps/_workspace/src/github.com/ugorji/go/codec/gen-helper.generated.go
… Godeps/_workspace/src/github.com/ugorji/go/codec/gen-helper.go.tmpl
… Godeps/_workspace/src/github.com/ugorji/go/codec/gen.generated.go
… Godeps/_workspace/src/github.com/ugorji/go/codec/gen.go
… Godeps/_workspace/src/github.com/ugorji/go/codec/helper.go
… Godeps/_workspace/src/github.com/ugorji/go/codec/helper_internal.go
… Godeps/_workspace/src/github.com/ugorji/go/codec/helper_not_unsafe.go
… Godeps/_workspace/src/github.com/ugorji/go/codec/helper_test.go
… Godeps/_workspace/src/github.com/ugorji/go/codec/helper_unsafe.go
… Godeps/_workspace/src/github.com/ugorji/go/codec/json.go
… Godeps/_workspace/src/github.com/ugorji/go/codec/msgpack.go
… Godeps/_workspace/src/github.com/ugorji/go/codec/noop.go
… Godeps/_workspace/src/github.com/ugorji/go/codec/prebuild.sh
… Godeps/_workspace/src/github.com/ugorji/go/codec/py_test.go
… Godeps/_workspace/src/github.com/ugorji/go/codec/rpc.go
… Godeps/_workspace/src/github.com/ugorji/go/codec/simple.go
… Godeps/_workspace/src/github.com/ugorji/go/codec/test-cbor-goldens.json