This change speeds up response unmarshal ~2x:

BEFORE:

```
BenchmarkSmallResponseUnmarshal	  10000	   164524 ns/op
BenchmarkManySmallResponseUnmarshal	    100	 13916636 ns/op
BenchmarkMediumResponseUnmarshal	   1000	  1974295 ns/op
BenchmarkLargeResponseUnmarshal	     20	 80462001 ns/op
```

AFTER:

```
BenchmarkSmallResponseUnmarshal	   20000	     75243 ns/op
BenchmarkManySmallResponseUnmarshal	     200	   6629661 ns/op
BenchmarkMediumResponseUnmarshal	    1000	   1359041 ns/op
BenchmarkLargeResponseUnmarshal	      20	  61600978 ns/op
```

Its performance is nearly the same as the one benchmarked in https://github.com/coreos/go-etcd/pull/218. So two packages perform similarly.

@wojtek-t 
I have attempted to use ffjson, but it doesn't perform as well as codec. Here is its benchmark:
```
BenchmarkSmallResponseUnmarshal	  10000	   117718 ns/op
BenchmarkManySmallResponseUnmarshal	    100	 10459107 ns/op
BenchmarkMediumResponseUnmarshal	   1000	  1659496 ns/op
BenchmarkLargeResponseUnmarshal	     20	 74711133 ns/op
```
So I still use codec in etcd/client pkg.

/cc @fgrzadkowski @wojtek-t @timothysc

fixes #3193 
ref https://github.com/kubernetes/kubernetes/issues/11962