@wojtek-t 2015-08-17
Thanks for comparison of ffjson & ugorij - this is really useful.

@fgrzadkowski - since you added that codec to the previous client, can you please take a look?