@wojtek-t 2015-08-17
Since @fgrzadkowski is OOO, I think it doesn't make sence to block on him.

I looked into it and it LGTM.