@wojtek-t 2015-09-01
Yes - I think it would be great if you could merge it soon - I will be trying to use new etcd client in kubernetes soon.