@timothysc 2015-09-03
/cc @danmcp - Fast unmarshaling which will also feed through the updated client.